requirements.txt
asgiref==3.8.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.1
django-allauth==0.63.3
django-cors-headers==4.4.0
django-softdelete==0.10.5
djangorestframework==3.15.2
drf-spectacular==0.27.2
gunicorn==22.0.0
h11==0.14.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-decouple==3.8
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
whitenoise==6.7.0